Fix flaky test testSendCorruptBytesToReplica
This test is failing with WindowsFS.
This change moves SegRep corruption tests to a separate IT class and skips running on WindowsFS.
